+++ stream-compositions/services/product-composition-service/src/main/java/com/backbase/stream/compositions/product/core/service/impl/ArrangementPostIngestionServiceImpl.java
@@ -98,7 +98,7 @@ public class ArrangementPostIngestionServiceImpl implements ArrangementPostInges
 
     private TransactionPullIngestionRequest buildTransactionPullRequest(ArrangementIngestResponse res) {
         return new TransactionPullIngestionRequest()
-                .withArrangementId(res.getArrangementInternalId())
-                .withExternalArrangementId(res.getArrangement().getExternalArrangementId());
+                .arrangementId(res.getArrangementInternalId())
+                .externalArrangementId(res.getArrangement().getExternalArrangementId());
     }
 }