+++ stream-legal-entity/legal-entity-http/src/main/java/com/backbase/stream/config/LegalEntityHttpConfiguration.java
@@ -1,14 +1,15 @@
 package com.backbase.stream.config;
 
-import org.springframework.boot.actuate.trace.http.HttpTraceRepository;
-import org.springframework.boot.actuate.trace.http.InMemoryHttpTraceRepository;
+import org.springframework.boot.actuate.web.exchanges.InMemoryHttpExchangeRepository;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.web.server.ServerHttpSecurity;
 import org.springframework.security.web.server.SecurityWebFilterChain;
+import org.springframework.validation.annotation.Validated;
 
 @Configuration
+@Validated
 public class LegalEntityHttpConfiguration {
 
     @Bean
@@ -29,8 +30,8 @@ public class LegalEntityHttpConfiguration {
      */
     @Bean
     @ConditionalOnExpression("${management.endpoints.enabled-by-default:false} or ${management.trace.http.enabled:false}")
-    public HttpTraceRepository httpTraceRepository() {
-        return new InMemoryHttpTraceRepository();
+    public InMemoryHttpExchangeRepository httpTraceRepository() {
+        return new InMemoryHttpExchangeRepository();
     }
 
 }